Initial commit of animationChecker script.
AnimationChecker is a PowerShell script that checks Captivate generated
SCORM packages for HTML5 CreateJS web objects. It will generate a report
of any objects that do not have the CreateJS libraries included in the
package. If this library is missing, the object will fail to function on
computers without internet access.